https://user-images.githubusercontent.com/108204659/178368662-c5484e21-d8e3-42b1-bf3c-d3489d82b79f.png
I personally have been impeded from getting a job and during an extended period of time frwhere the economic impact would also be my own tax dollars to benefit the local, state and federal government - and by a pack of deplorable people who all will go to prison for the things they have done. With all due respect the culprits are all confidential - until they are not. Just like the examples above and as seen in other representations of fugitives in American Greed, to catch a predator, notwithstanding the success stories. A good example of one that I enjoy watching, despite those losses is this video https://www.youtube.com/watch?v=i1C1FLlICXs i in fact moved into that building (590 Madison Avenue, IBM building) for $75 million dollars into an office that was PRE-WIRED by the FBI. Despite the severity and scrutiny as implied by their abilities in the video above, had no issues condeucting my business and work inside of my office - because I am not a felon and do not violate securities law and would not in a million years even consider the though of a deplorable breach of the law or conduct at any point. Which bring me back to the INEVITABLE explanations and conversation in the case of the $75 million dollar scheme headed by Raj Rajaratnam and the 46 suspects and 26 who simply pleaded guilty to insider trading (https://www.youtube.com/watch?v=JNLmlkyda58).

this generally also follows the individuals and their " grandeur " in the beginning of the movie - end up in a cell or dead by the end of the movie.

WILL END UP LIKE THEY HIT BY THE RAPID TRANSIT AUTHORITY AFTER
☆ ☆ ☆ ☆ ☆ 
https://www.youtube.com/watch?v=pzhD2MyeIbY&t=1940




This generally also follows the individuals and principles who avoid prosecution in their self-declared state of " grandeur ", exmplified in the beginning of the movie - and demonstrably without the intervention of Federal authorities end up in a yellow-jumpsuit and/or cell (or dead) by the end of the movie, for failing to abide by the laws - and are left with their children who end up sounding like they also got hit by the rapid-transit-authority by the end of the movie - without any exemption having violated that rap sheet, as referenced previously.

☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ 

OG SOURCE >> https://www.fbi.gov/video-repository/newss-chasing-the-dragon-the-life-of-an-opiate-addict/file_view 

https://www.youtube.com/watch?v=pzhD2MyeIbY&t=1940




☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ ☆ 